+++ molvis/vasp/python/vasputil.py
@@ -91,11 +91,9 @@ def parseFile(file, flipSign=False, centerData=True):
         for c in line.strip().split():
             chg.append(sign * float(c))
 
-
     chgdata = numpy.array(chg, dtype=numpy.float32)
     volume = numpy.dot(basis[0], numpy.cross(basis[1], basis[2]))
     chgdata = chgdata / numpy.float32(volume)
-
 
     chgdata.shape = (dims[2], dims[1], dims[0])
 
@@ -117,8 +115,8 @@ def parseFile(file, flipSign=False, centerData=True):
     volume.dataMap.valueAxis.name = "Charge Density"
     volume.dataMap.valueAxis.unit = ivw.data.Unit("e/Angstrom^3")
 
-    volume.basis = ivw.glm.mat3(basis)
-    volume.offset = ivw.glm.vec3(offset)
+    volume.basis = ivw.glm.dmat3(basis)
+    volume.offset = ivw.glm.dvec3(offset)
 
     atoms = []
     for i, n in enumerate(nelem):
